requirements.txt
streamlit
langchain
langchain-ollama
langchain-openai
langchain-pinecone
pinecone
python-dotenv
python-docx
pypdf
langgraph
langchain-community